Rename "_legacy" dir to "legacy"
