+++ packages/compose/src/hooks/use-resize-observer/legacy/index.native.js
